Initial JSS integration via JSSImporter.py commit
With separate recipe(VLC is what I happened to test with,) this
processor can currently find categories and pkgs by product and
prod_version names respectively, and if absent creates them in the JSS
and does the filedrop to a locally path'd distribution point. The code
borrows heavily from TVSutton and acidprime's work on the DeployStudio
REST interface, uses generous spoonfuls of MunkiImporter's framework,
borrows the salient part of the Copier class's copy method. TODO:
policy and group interaction