As per @timsutton's feedback, allow optout as Pref
This alters the main function to allow for code signature verification
to be overwritten by a global pref or a key fed to the running of a
recipe. It merely checks for the existence of ANY value on that key,
and returns from main (skipping the actual check code) as a result.