Improved default pkg_path
Currently if a glob character is used (like *.pkg) and a default pkg_path isn't defined, then the pkg_path will be "*.pkg".

This change switches the default from source_pkg to matched_source_path.  This way if glob character are used the pkg_path will use the actual file name that was matched and not the glob pattern.