Introduce the preprocessor MunkiNormalizePath
MunkiNormalizePath is aim to be used with --pre to standardize imported item name and subpath
It will update NAME and MUNKI_REPO_SUBDIR variable to replace all non alphanumeric char with an underscore and will also convert everything to lowercase

Using it shouldn't break correclty written recipes. However, some has been written to use NAME everywhere, even on non-varaible things.

When deploying this preprocessor, be sure to check that all your recipes are still working. If there is any issue, send a PR to the recipe maintainer to avoid using NAME for non variable things.