generate_processor_docs with flake8+isort (#517)
* generate_processor_docs with flake8+isort

* Also change ''' to """

* Fixing the import dependencies being broken by isort

The import order in this case is dependent on loading files into the path, and using imp to load a source file directly.

To address this, the imports are wrapped in a try/except and will simply exit 1 if they don't function. This allows the order dependencies to exist without letting isort break them.