Update FileFinder.py
Sets recursive to True for glob.

This will only take affect if the "**" pattern is passed to FileFinder, exisiting recipes etc will be fine (& have tested a few post change).

'''
If recursive is true, the pattern “**” will match any files and zero or more directories, subdirectories and symbolic links to directories. If the pattern is followed by an os.sep or os.altsep then files will not match.
'''