Force normalized (LF) line endings for all files.
Windows support introduces the possibility of files having different
line ending conventions. There are already examples of this in the
repository despite the macOS specific nature of the tool.

This commit fixes all existing instances of CRLF endings, and ensures
that no inconsistencies can be accidentally introduced later. Should the
need arise it will still be possible to commit a file with CRLF.