Tooling updates: black, isort, flake8
isort v4.x doesn't have the ability to inline control certain sorting
constructs, necessary for making tooling that's less sensitive to its
execution path. flake8 is updated since it seems likely it was going to
be updated soon given that with this commit 2/3 of the pre-commit hooks
would have been updated.