Resource and processor loading uses `pkg_resources`
Change the way autopkg discovers built-in processors, and its version to
use the `pkg_resources` module instead of raw filesystem access. This is
a step towards allowing autopkg to be bundled into a single-file
executable, such as with the `zipapp` module. The value being that
single-file deployment is much easier and more reliable than a raw
source layout.