Remove use of site exit in autopkg main
The `site` module fiddles with the global namespace to add `exit`.
However, this is only meant for interactive use. Generally speaking,
this module is auto-imported so it was never an issue. However, in some
environments the module is not present, or the `-S` flag is used to
disable importing it.