AppPkgCreator: New Optional Input: version_key
Adding a feature to allow an alternate version key in `Info.plist`. The `.get` method for the `dict` class allows NoneType arguments without throwing a TypeError exception. So if `version_key` is unset, it will return `None`, which will in turn return `None` on get `.get` for `infoplist`, defaulting to `CFBundleShortVersionString`.

This will not affect existing recipes.