Update xattr.py
xattr does not work in some Linux file systems, or in Linux running under WSL, or in some docker containers in general. This will make the failures of xattr no longer hard errors in these cases.

Use of xattr for storing etag data is not ideal in the first place and is already solved in URLDownloaderPython by using a json metadata file instead, but this will help with compatibility in general for older methods. This will have the drawback of not detecting that the already downloaded file is the same correctly due to lack of metadata for any recipe currently relying on xattr. That said, redownloading the file seems better than throwing hard errors.