Fix curl command error in URLDownloader.prefetch_filename()
Previously, combining '--head' and '--request GET' allowed curl to perform a
GET request and stop after receiving the headers. This workaround was added to
handle servers that return HTTP Forbidden or incorrect headers for a HEAD
request. However, with the update to curl 8.6.0 on macOS 14.5, this
combination results in an error ("curl: (8) Weird server reply").